Switching from mutagen to ffmpeg, implemented duplicate handling, and added a margin of error adjustment

+++ disc_filler.py
@@ -2,8 +2,7 @@ import argparse
 import random
 import shutil
 from pathlib import Path
-import mutagen
-from mutagen.mp3 import MP3
+import ffmpeg
 
 
 def main(source_dir, destination):
@@ -14,35 +13,52 @@ def main(source_dir, destination):
     :param destination: Path - directory to copy files to
     """
     file_list = list()
+    tracks_already_included = list()
     for file in source_dir.glob('**/*.mp3'):
         file_list.append(file)
 
     # 74 minute disc, LP4
     available_seconds = (74 * 60 * 4)
+    # tweak to try to overcome the error margin
+    available_seconds *= 1.013
 
     while available_seconds > 30:
         # Grab a random file and remove it from the list
         current_file = file_list[random.randint(0, (len(file_list)) - 1)]
+
         try:
-            duration = MP3(current_file).info.length
-            print(f"File: {current_file}\n"
-                  f"Time remaining: {available_seconds}\n"
-                  f"File duration: {duration}")
+            title_artist = f"{ffmpeg.probe(current_file)['format']['tags']['title']} " \
+                           f"- {ffmpeg.probe(current_file)['format']['tags']['artist']}"
+        except ffmpeg.Error:
+            print(f"\n{current_file} looks weird, skipping it.\n")
+            continue
 
-            if available_seconds > duration:
-                # copy the file to the output dir
-                available_seconds -= int(duration)
+        if title_artist in tracks_already_included:
+            print("\nTrack already included, skipping\n")
+            continue
+        else:
+            try:
+                duration = float(ffmpeg.probe(current_file)['format']['duration'])
+                print(f"File: {current_file}\n"
+                      f"Time remaining: {available_seconds}\n"
+                      f"File duration: {duration}")
 
-                randomise_filename = True
-                if randomise_filename:
-                    output_filename = Path(str(destination) + f"/{str(random.randint(0, 99))} " + current_file.name)
-                else:
-                    output_filename = Path(str(destination) + f"/" + current_file.name)
-                shutil.copyfile(current_file, output_filename)
+                if available_seconds > duration:
+                    # copy the file to the output dir
+                    available_seconds -= duration
 
-        except mutagen.mp3.HeaderNotFoundError:
-            # This MP3 file has some sort of error, skip it
-            pass
+                    randomise_filename = True
+                    if randomise_filename:
+                        output_filename = Path(str(destination) + f"/{str(random.randint(0, 99))} " + current_file.name)
+                    else:
+                        output_filename = Path(str(destination) + f"/" + current_file.name)
+                    # copy the file to the output directory
+                    shutil.copyfile(current_file, output_filename)
+
+                    # add it to the list of tracks included to avoid duplicates
+                    tracks_already_included.append(title_artist)
+            except ffmpeg.Error:
+                continue
 
 
 if __name__ == '__main__':
